realtek: 6.12: refactor get_mac_eee
Upstream will get rid of the get_mac_eee() function in the DSA
driver and replace it by a boolean alternative. While we fill a
lot of data here other drivers only return if EEE is available
or not for a port. To make the next kernel upgrade easier follow
that design.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>